refactor: deduplicate helpers into ffmpeg_helpers.py, fix pre-existing test failures
- Move _seconds_to_srt_time() from ai_engine.py and engine.py to ffmpeg_helpers.py
- Move _run_ffprobe_json() from engine.py to ffmpeg_helpers.py (now uses shared _run_ffmpeg)
- Rewrite ai_engine._run_ffprobe() to use _run_ffmpeg instead of raw subprocess
- Fix client.add_generated_audio() passing wrong kwarg names
- Fix test_version_flag using hardcoded version instead of __version__
